Fix notification sitemap sub-page navigation (#842)
When a notification action targets a specific sitemap widget via
`ui:/basicui/app?w=<pageId>&sitemap=<name>`, the app was navigating
to the root sitemap page instead of the intended sub-page. This was
caused by a race condition between SwiftUI's `.task` (which fired
`startPageHandling` with an empty `pageId`) and the async
`pushSitemap` call that set the correct `pageId` afterwards.

Fix by extracting `HostingSitemapViewController` into its own file
and rewriting `pushSitemap(name:path:)` to construct a new
`HostingSitemapViewController` pre-configured with the target page
URL and page ID, then push it directly onto the navigation stack.
This eliminates the race: the view model already has the correct
`pageId` before the view appears and `startPageHandling` fires.

Also adds `SitemapPageViewModel.init(sitemapName:pageUrl:title:pageId:)`
and `configureSitemap(name:pageId:)` to support this pattern.
