+++ openHAB/UI/OpenHABRootViewController.swift
@@ -69,78 +69,6 @@ protocol ModalHandler: AnyObject {
     func modalDismissed(to: TargetController)
 }
 
-class HostingSitemapViewController: UIHostingController<SitemapNavigationView>, OpenHABViewable {
-    private let viewModel: SitemapPageViewModel
-
-    private weak var rootViewController: OpenHABRootViewController?
-
-    init() {
-        let viewModel = SitemapPageViewModel()
-        self.viewModel = viewModel
-        super.init(rootView: SitemapNavigationView(viewModel: viewModel) {})
-    }
-
-    @available(*, unavailable)
-    @objc dynamic required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Hide UIKit navigation bar since SwiftUI now handles navigation
-        navigationController?.setNavigationBarHidden(true, animated: false)
-    }
-
-    override func viewWillAppear(_ animated: Bool) {
-        super.viewWillAppear(animated)
-        // Ensure UIKit navigation bar stays hidden when transitioning from other views
-        navigationController?.setNavigationBarHidden(true, animated: animated)
-    }
-
-    func setRootViewController(_ rootViewController: OpenHABRootViewController) {
-        self.rootViewController = rootViewController
-        // Update the closure after initialization
-        rootView = SitemapNavigationView(viewModel: viewModel) { [weak self] in
-            self?.rootViewController?.showSideMenu()
-        }
-    }
-
-    func getSitemapTitle() -> String {
-        viewModel.pageTitle
-    }
-
-    func viewName() -> String { "sitemap" }
-
-    nonisolated func reloadView() {
-        Task { @MainActor in
-            await viewModel.reload()
-        }
-    }
-
-    func pushSitemap(name: String, path: String?) async {
-        // Implement pushing logic into SitemapPageViewModel
-        await viewModel.pushSitemap(name: name, path: path)
-    }
-
-    @MainActor
-    func refreshOnForegroundIfNeeded() {
-        // Avoid restarting the very first page load on initial app activation.
-        guard viewModel.currentPage != nil || !viewModel.isLoading else { return }
-        viewModel.refreshOnForeground()
-    }
-
-    // swiftlint:disable:next  function_parameter_count
-    func showPopupMessage(seconds: Double,
-                          title: String,
-                          message: String,
-                          theme: Theme,
-                          viewTapAction: (() -> Void)?,
-                          buttonTitle: String,
-                          buttonAction: (() -> Void)?) {}
-
-    func hidePopupMessages() {}
-}
-
 // MARK: - Search Controller Delegates
 
 // swiftlint:disable type_body_length
